ai-server/requirements.txt
tensorflow==2.10.0
numpy==1.23.5
flask==2.2.3
flask-cors==3.0.10
gunicorn==20.1.0
pillow==9.3.0
opencv-python==4.6.0.66
insightface==0.6.2
facenet-pytorch==2.5.2
onnxruntime==1.13.1
scikit-learn==1.0.2
pydantic==1.10.7
python-dotenv==1.0.0
requests==2.28.2
redis==4.5.1